requirements.txt
python-dotenv==0.10.1
flask==1.0.2
gunicorn==19.9.0
Flask-WTF==0.14.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
flasgger==0.9.2
requests==2.21.0
Flask-Cors==3.0.8
Flask-Migrate==2.4.0
mysqlclient==1.4.2.post1